rename toool definition
